feat: Add custom rectangle view
