+++ DrawingApp/DrawingApp/View/RectangleView.swift
@@ -1,0 +1,24 @@
+//
+//  RectangleView.swift
+//  DrawingApp
+//
+//  Created by 송태환 on 2022/03/07.
+//
+
+import UIKit
+
+class RectangleView: UIView {
+    func setBackgroundColor(color: Color, alpha: Alpha = .opaque) {
+        let alphaValue = alpha.convert(using: CGFloat.self) / 10
+        self.backgroundColor = UIColor(red: color.red, green: color.green, blue: color.blue, alpha: alphaValue)
+    }
+    
+    func setBackgroundColor(with alpha: CGFloat) {
+        let color = self.backgroundColor?.withAlphaComponent(alpha)
+        self.backgroundColor = color
+    }
+    
+    func setAlpha(alpha: Alpha) {
+        self.alpha = alpha.convert(using: CGFloat.self) / 10
+    }
+}
